requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
category-encoders==2.6.2
Django==4.2.5
djangorestframework==3.14.0
importlib-resources==6.0.1
joblib==1.3.2
numpy==1.24.4
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
statsmodels==0.14.0
threadpoolctl==3.2.0
typing-extensions==4.8.0
tzdata==2023.3
zipp==3.17.0
